LayoutTests/imported/w3c/web-platform-tests/trusted-types/DOMWindowTimers-setTimeout-setInterval-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.


Harness Error (TIMEOUT), message = null

PASS window.setTimeout assigned via policy (successful Script transformation).
PASS window.setInterval assigned via policy (successful Script transformation).
TIMEOUT window.setTimeout assigned via default policy (successful Script transformation). Test timed out
TIMEOUT window.setInterval assigned via default policy (successful Script transformation). Test timed out
PASS window.setTimeout assigned with a function handler shouldn't go through default policy.
PASS window.setInterval assigned with a function handler shouldn't go through default policy.
